created stable controllers for all models

+++ src/main/java/com/cailou/esigAtividadeTecnica/repository/TokenRepository.java
@@ -3,7 +3,5 @@ package com.cailou.esigAtividadeTecnica.repository;
 import com.cailou.esigAtividadeTecnica.model.TokenModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface TokenRepository extends JpaRepository<TokenModel, BigInteger> {
+public interface TokenRepository extends JpaRepository<TokenModel, Long> {
 }
